+++ bridge-adapters/kinetic-bridgehub-adapter-kinetic-platform/changelog.md
@@ -1,3 +1,6 @@
+Kinetic Platform Adapter (2022-12-23)
+* Added webApi search/retrieve functionality (space and kapp)
+
 Kinetic Platform Adapter (2020-10-26)
  * updated log4j and junit dependencies.  Junit went to 4.13.1 and log4j core api and slf4j-impl went to 2.13.2.  These we security upgrades for dependencies.
  * bumped  patch number